requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.30.1=pypi_0
accelerator=2024.3.8.dev1=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
bitsandbytes=0.42.0=pypi_0
blas=1.0=mkl
blis=0.7.11=pypi_0
bottle=0.12.25=pypi_0
bottleneck=1.3.7=py310ha9d4c09_0
brotli-python=1.0.9=py310h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
catalogue=2.0.10=pypi_0
certifi=2024.2.2=py310h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
confection=0.1.4=pypi_0
contourpy=1.2.1=pypi_0
cuda=11.8.0=0
cuda-cccl=11.8.89=0
cuda-command-line-tools=11.8.0=0
cuda-compiler=11.8.0=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cuobjdump=11.8.86=0
cuda-cupti=11.8.87=0
cuda-cuxxfilt=11.8.86=0
cuda-demo-suite=11.8.86=0
cuda-documentation=11.8.86=0
cuda-driver-dev=11.8.89=0
cuda-gdb=11.8.86=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.8.89=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.8.86=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.8.86=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-nvvp=11.8.87=0
cuda-profiler-api=11.8.86=0
cuda-runtime=11.8.0=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.8.0=0
cuda-tools=11.8.0=0
cuda-visual-tools=11.8.0=0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
dgl=2.2.1+cu118=pypi_0
einops=0.8.0=pypi_0
en-core-web-md=3.7.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
fonttools=4.53.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
gds-tools=1.4.0.31=0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.23.0=pypi_0
idna=3.7=py310h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.3=py310h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
langcodes=3.4.0=pypi_0
language-data=1.2.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
marisa-trie=1.1.1=pypi_0
markupsafe=2.1.3=py310h5eee18b_0
matplotlib=3.9.0=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
multidict=6.0.5=pypi_0
murmurhash=1.0.10=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
nltk=3.8.1=pypi_0
nsight-compute=2022.3.0.22=0
numexpr=2.8.4=py310h8879344_0
numpy=1.24.3=py310hd5efca6_0
numpy-base=1.24.3=py310h8e6c178_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
packaging=24.0=pypi_0
pandas=2.2.1=py310h6a678d5_0
pillow=10.3.0=py310h5eee18b_0
pip=24.0=py310h06a4308_0
preshed=3.0.9=pypi_0
psutil=5.9.8=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pyg-lib=0.4.0+pt22cu118=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.0=h12debd9_5
python-dateutil=2.9.0post0=py310h06a4308_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.2.2=py3.10_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2024.1=py310h06a4308_0
pyyaml=6.0.1=py310h5eee18b_0
readline=8.2=h5eee18b_0
regex=2024.5.15=pypi_0
requests=2.31.0=py310h06a4308_1
safetensors=0.4.3=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.0=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.2.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smart-open=6.4.0=pypi_0
sortednp=0.5.0=pypi_0
spacy=3.7.4=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
spacy-pkuseg=0.0.33=pypi_0
sqlite=3.45.3=h5eee18b_0
srsly=2.4.8=pypi_0
sympy=1.12=py310h06a4308_0
thinc=8.2.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
torch-cluster=1.6.3+pt22cu118=pypi_0
torch-geometric=2.5.3=pypi_0
torch-scatter=2.1.2+pt22cu118=pypi_0
torch-sparse=0.6.18+pt22cu118=pypi_0
torch-spline-conv=1.2.2+pt22cu118=pypi_0
torchaudio=2.2.2=py310_cu118
torchdata=0.7.1=pypi_0
torchtriton=2.2.0=py310
torchvision=0.17.2=py310_cu118
tqdm=4.66.4=pypi_0
transformers=4.41.0=pypi_0
typer=0.9.4=pypi_0
typing_extensions=4.11.0=py310h06a4308_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.1=py310h06a4308_0
waitress=3.0.0=pypi_0
wasabi=1.1.2=pypi_0
weasel=0.3.4=pypi_0
wheel=0.43.0=py310h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zh-core-web-md=3.7.0=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
